split object_list.rs into several files
