+++ src/cwe_checker_lib/src/analysis/pointer_inference/object_list/mod.rs
@@ -5,6 +5,10 @@ use crate::{abstract_domain::*, utils::binary::RuntimeMemoryImage};
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
 
+mod cwe_helpers;
+mod list_manipulation;
+mod id_manipulation;
+
 /// The list of all known abstract objects.
 ///
 /// Each abstract object is unique in the sense that there is exactly one abstract identifier pointing to it.
@@ -41,109 +45,6 @@ impl AbstractObjectList {
             ),
         );
         AbstractObjectList { objects }
-    }
-
-    /// Check the state of a memory object at a given address.
-    /// Returns `true` if at least one of the targets of the pointer is dangling.
-    /// If `report_unknown_states` is `true`,
-    /// then objects with unknown states get reported if they are unique.
-    /// I.e. objects representing more than one actual object (e.g. an array of object) will not get reported,
-    /// even if their state is unknown and `report_unknown_states` is `true`.
-    pub fn is_dangling_pointer(&self, address: &Data, report_unknown_states: bool) -> bool {
-        for id in address.referenced_ids() {
-            if let Some((object, _offset_id)) = self.objects.get(id) {
-                match (report_unknown_states, object.get_state()) {
-                    (_, ObjectState::Dangling) => return true,
-                    (true, ObjectState::Unknown) => {
-                        if object.is_unique() {
-                            return true;
-                        }
-                    }
-                    _ => (),
-                }
-            }
-        }
-        // No dangling pointer found
-        false
-    }
-
-    /// Mark all memory objects targeted by the given `address` pointer,
-    /// whose state is either dangling or unknown,
-    /// as flagged.
-    pub fn mark_dangling_pointer_targets_as_flagged(&mut self, address: &Data) {
-        for id in address.referenced_ids() {
-            let (object, _) = self.objects.get_mut(id).unwrap();
-            if matches!(
-                object.get_state(),
-                ObjectState::Unknown | ObjectState::Dangling
-            ) {
-                object.set_state(ObjectState::Flagged);
-            }
-        }
-    }
-
-    /// Check whether a memory access at the given address (and accessing `size` many bytes)
-    /// may be an out-of-bounds memory access.
-    ///
-    /// Note that `Top` values as addresses are not marked as out-of-bounds,
-    /// since they are more likely due to analysis imprecision than to actual out-of-bounds access.
-    pub fn is_out_of_bounds_mem_access(
-        &self,
-        address: &Data,
-        size: ByteSize,
-        global_data: &RuntimeMemoryImage,
-    ) -> bool {
-        // TODO: Check the interval even if the Top-flag in the DataDomain is set!
-        if let Some(value) = address.get_absolute_value() {
-            if let Ok((start, end)) = value.try_to_offset_interval() {
-                if start < 0 || end < start {
-                    return true;
-                }
-                if global_data
-                    .is_interval_readable(start as u64, end as u64 + u64::from(size) - 1)
-                    .is_err()
-                {
-                    return true;
-                }
-            }
-        }
-        for (id, offset) in address.get_relative_values() {
-            if let Some((object, base_offset)) = self.objects.get(id) {
-                let adjusted_offset = offset.clone() + base_offset.clone();
-                if !adjusted_offset.is_top()
-                    && !object.access_contained_in_bounds(&adjusted_offset, size)
-                {
-                    return true;
-                }
-            }
-        }
-        false
-    }
-
-    /// Set the lower index bound for indices to be considered inside the memory object.
-    /// The bound is inclusive, i.e. the bound index itself is also considered to be inside the memory object.
-    ///
-    /// Any `bound` value other than a constant bitvector is interpreted as the memory object not having a lower bound.
-    pub fn set_lower_index_bound(&mut self, object_id: &AbstractIdentifier, bound: &ValueDomain) {
-        let (object, base_offset) = self.objects.get_mut(object_id).unwrap();
-        let bound = (bound.clone() + base_offset.clone())
-            .try_to_bitvec()
-            .map(|bitvec| bitvec.into())
-            .unwrap_or_else(|_| BitvectorDomain::new_top(bound.bytesize()));
-        object.set_lower_index_bound(bound);
-    }
-
-    /// Set the upper index bound for indices to be considered inside the memory object.
-    /// The bound is inclusive, i.e. the bound index itself is also considered to be inside the memory object.
-    ///
-    /// Any `bound` value other than a constant bitvector is interpreted as the memory object not having an upper bound.
-    pub fn set_upper_index_bound(&mut self, object_id: &AbstractIdentifier, bound: &ValueDomain) {
-        let (object, base_offset) = self.objects.get_mut(object_id).unwrap();
-        let bound = (bound.clone() + base_offset.clone())
-            .try_to_bitvec()
-            .map(|bitvec| bitvec.into())
-            .unwrap_or_else(|_| BitvectorDomain::new_top(bound.bytesize()));
-        object.set_upper_index_bound(bound);
     }
 
     /// Get the value at a given address.
@@ -199,128 +100,6 @@ impl AbstractObjectList {
         }
     }
 
-    /// Replace one abstract identifier with another one. Adjust offsets of all pointers accordingly.
-    ///
-    /// **Example:**
-    /// Assume the `old_id` points to offset 0 in the corresponding memory object and the `new_id` points to offset -32.
-    /// Then the offset_adjustment is -32.
-    /// The offset_adjustment gets *added* to the base offset in `self.memory.ids` (so that it points to offset -32 in the memory object),
-    /// while it gets *subtracted* from all pointer values (so that they still point to the same spot in the corresponding memory object).
-    pub fn replace_abstract_id(
-        &mut self,
-        old_id: &AbstractIdentifier,
-        new_id: &AbstractIdentifier,
-        offset_adjustment: &ValueDomain,
-    ) {
-        let negative_offset = -offset_adjustment.clone();
-        for (object, _) in self.objects.values_mut() {
-            object.replace_abstract_id(old_id, new_id, &negative_offset);
-        }
-        if let Some((object, old_offset)) = self.objects.remove(old_id) {
-            let new_offset = old_offset + offset_adjustment.clone();
-            self.objects.insert(new_id.clone(), (object, new_offset));
-        }
-    }
-
-    /// Remove the memory object that `object_id` points to from the object list.
-    pub fn remove_object(&mut self, object_id: &AbstractIdentifier) {
-        self.objects.remove(object_id);
-    }
-
-    /// Add a new abstract object to the object list
-    ///
-    /// If an object with the same ID already exists,
-    /// the object is marked as non-unique and merged with the newly created object.
-    pub fn add_abstract_object(
-        &mut self,
-        object_id: AbstractIdentifier,
-        initial_offset: ValueDomain,
-        type_: ObjectType,
-        address_bytesize: ByteSize,
-    ) {
-        let new_object = AbstractObject::new(type_, address_bytesize);
-        if let Some((object, offset)) = self.objects.get_mut(&object_id) {
-            // If the identifier already exists, we have to assume that more than one object may be referenced by this identifier.
-            object.mark_as_not_unique();
-            *object = object.merge(&new_object);
-            *offset = offset.merge(&initial_offset);
-        } else {
-            self.objects.insert(object_id, (new_object, initial_offset));
-        }
-    }
-
-    /// Return all IDs that may be referenced by the memory object pointed to by the given ID.
-    /// The returned set is an overapproximation of the actual referenced IDs.
-    pub fn get_referenced_ids_overapproximation(
-        &self,
-        id: &AbstractIdentifier,
-    ) -> BTreeSet<AbstractIdentifier> {
-        if let Some((object, _offset)) = self.objects.get(id) {
-            object.get_referenced_ids_overapproximation().clone()
-        } else {
-            BTreeSet::new()
-        }
-    }
-
-    /// Return all IDs that get referenced by the memory object pointed to by the given ID.
-    /// The returned set is an underapproximation of the actual referenced IDs,
-    /// since only still tracked pointers inside the memory object are used to compute it.
-    pub fn get_referenced_ids_underapproximation(
-        &self,
-        id: &AbstractIdentifier,
-    ) -> BTreeSet<AbstractIdentifier> {
-        if let Some((object, _offset)) = self.objects.get(id) {
-            object.get_referenced_ids_underapproximation()
-        } else {
-            panic!("Abstract ID not associated to an object")
-        }
-    }
-
-    /// For abstract IDs not contained in the provided set of IDs
-    /// remove the corresponding abstract objects.
-    ///
-    /// This function does not remove any pointer targets in the contained abstract objects.
-    pub fn remove_unused_objects(&mut self, ids_to_keep: &BTreeSet<AbstractIdentifier>) {
-        let all_ids: BTreeSet<AbstractIdentifier> = self.objects.keys().cloned().collect();
-        let ids_to_remove = all_ids.difference(ids_to_keep);
-        for id in ids_to_remove {
-            self.objects.remove(id);
-        }
-    }
-
-    /// Get all object IDs.
-    pub fn get_all_object_ids(&self) -> BTreeSet<AbstractIdentifier> {
-        self.objects.keys().cloned().collect()
-    }
-
-    /// Mark a memory object as already freed (i.e. pointers to it are dangling).
-    ///
-    /// If the object cannot be identified uniquely, all possible targets are marked as having an unknown status.
-    /// Returns either a non-empty list of detected errors (like possible double frees) or `OK(())` if no errors were found.
-    pub fn mark_mem_object_as_freed(
-        &mut self,
-        object_pointer: &Data,
-    ) -> Result<(), Vec<(AbstractIdentifier, Error)>> {
-        let ids: Vec<AbstractIdentifier> = object_pointer.referenced_ids().cloned().collect();
-        let mut possible_double_free_ids = Vec::new();
-        if ids.len() > 1 {
-            for id in ids {
-                if let Err(error) = self.objects.get_mut(&id).unwrap().0.mark_as_maybe_freed() {
-                    possible_double_free_ids.push((id.clone(), error));
-                }
-            }
-        } else if let Some(id) = ids.get(0) {
-            if let Err(error) = self.objects.get_mut(id).unwrap().0.mark_as_freed() {
-                possible_double_free_ids.push((id.clone(), error));
-            }
-        }
-        if possible_double_free_ids.is_empty() {
-            Ok(())
-        } else {
-            Err(possible_double_free_ids)
-        }
-    }
-
     /// Assume that arbitrary writes happened to a memory object,
     /// including adding pointers to targets contained in `new_possible_reference_targets` to it.
     ///
@@ -336,35 +115,6 @@ impl AbstractObjectList {
     ) {
         if let Some((object, _)) = self.objects.get_mut(object_id) {
             object.assume_arbitrary_writes(new_possible_reference_targets);
-        }
-    }
-
-    /// Get the number of objects that are currently tracked.
-    #[cfg(test)]
-    pub fn get_num_objects(&self) -> usize {
-        self.objects.len()
-    }
-
-    /// Append those objects from another object list, whose abstract IDs are not known to self.
-    pub fn append_unknown_objects(&mut self, other_object_list: &AbstractObjectList) {
-        for (id, (other_object, other_offset)) in other_object_list.objects.iter() {
-            if self.objects.get(id) == None {
-                self.objects
-                    .insert(id.clone(), (other_object.clone(), other_offset.clone()));
-            }
-        }
-    }
-
-    /// Remove the provided IDs as targets from all pointers in all objects.
-    /// Also remove the objects, that these IDs point to.
-    pub fn remove_ids(&mut self, ids_to_remove: &BTreeSet<AbstractIdentifier>) {
-        for id in ids_to_remove {
-            if self.objects.get(id).is_some() {
-                self.objects.remove(id);
-            }
-        }
-        for (object, _) in self.objects.values_mut() {
-            object.remove_ids(ids_to_remove);
         }
     }
 
@@ -437,164 +187,4 @@ impl AbstractObjectList {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn bv(value: i64) -> ValueDomain {
-        ValueDomain::from(Bitvector::from_i64(value))
-    }
-
-    fn new_id(name: &str) -> AbstractIdentifier {
-        AbstractIdentifier::new(
-            Tid::new("time0"),
-            AbstractLocation::Register(name.into(), ByteSize::new(8)),
-        )
-    }
-
-    #[test]
-    fn abstract_object_list() {
-        let mut obj_list =
-            AbstractObjectList::from_stack_id(new_id("RSP".into()), ByteSize::new(8));
-        assert_eq!(obj_list.objects.len(), 1);
-        assert_eq!(obj_list.objects.values().next().unwrap().1, bv(0));
-
-        let pointer = DataDomain::from_target(new_id("RSP".into()), bv(8));
-        obj_list.set_value(pointer.clone(), bv(42).into()).unwrap();
-        assert_eq!(
-            obj_list.get_value(&pointer, ByteSize::new(8)),
-            bv(42).into()
-        );
-
-        let mut other_obj_list =
-            AbstractObjectList::from_stack_id(new_id("RSP".into()), ByteSize::new(8));
-        let second_pointer = DataDomain::from_target(new_id("RSP".into()), bv(-8));
-        other_obj_list
-            .set_value(pointer.clone(), bv(42).into())
-            .unwrap();
-        other_obj_list
-            .set_value(second_pointer.clone(), bv(35).into())
-            .unwrap();
-        assert_eq!(
-            other_obj_list.get_value(&second_pointer, ByteSize::new(8)),
-            bv(35).into()
-        );
-
-        other_obj_list.add_abstract_object(
-            new_id("RAX".into()),
-            bv(0),
-            ObjectType::Heap,
-            ByteSize::new(8),
-        );
-        let heap_pointer = DataDomain::from_target(new_id("RAX".into()), bv(8));
-        other_obj_list
-            .set_value(heap_pointer.clone(), bv(3).into())
-            .unwrap();
-
-        let mut merged = obj_list.merge(&other_obj_list);
-        assert_eq!(merged.get_value(&pointer, ByteSize::new(8)), bv(42).into());
-
-        assert!(merged
-            .get_value(&second_pointer, ByteSize::new(8))
-            .contains_top());
-        assert_eq!(
-            merged.get_value(&heap_pointer, ByteSize::new(8)),
-            bv(3).into()
-        );
-        assert_eq!(merged.objects.len(), 2);
-
-        merged
-            .set_value(pointer.merge(&heap_pointer), bv(3).into())
-            .unwrap();
-        assert_eq!(
-            merged.get_value(&pointer, ByteSize::new(8)),
-            IntervalDomain::mock(3, 42).with_stride(39).into()
-        );
-        assert_eq!(
-            merged.get_value(&heap_pointer, ByteSize::new(8)),
-            bv(3).into()
-        );
-        assert_eq!(merged.objects.len(), 2);
-
-        other_obj_list
-            .set_value(pointer.clone(), heap_pointer.clone())
-            .unwrap();
-        assert_eq!(
-            other_obj_list
-                .get_referenced_ids_overapproximation(&new_id("RSP".into()))
-                .len(),
-            1
-        );
-        assert_eq!(
-            *other_obj_list
-                .get_referenced_ids_overapproximation(&new_id("RSP".into()))
-                .iter()
-                .next()
-                .unwrap(),
-            new_id("RAX".into())
-        );
-
-        let modified_heap_pointer = DataDomain::from_target(new_id("ID2".into()), bv(8));
-        other_obj_list.replace_abstract_id(&new_id("RAX".into()), &new_id("ID2".into()), &bv(0));
-        assert_eq!(
-            other_obj_list.get_value(&pointer, ByteSize::new(8)),
-            modified_heap_pointer.clone()
-        );
-        assert_eq!(other_obj_list.objects.get(&new_id("RAX".into())), None);
-        assert!(matches!(
-            other_obj_list.objects.get(&new_id("ID2".into())),
-            Some(_)
-        ));
-
-        let mut ids_to_keep = BTreeSet::new();
-        ids_to_keep.insert(new_id("ID2".into()));
-        other_obj_list.remove_unused_objects(&ids_to_keep);
-        assert_eq!(other_obj_list.objects.len(), 1);
-        assert_eq!(
-            other_obj_list.objects.iter().next().unwrap().0,
-            &new_id("ID2".into())
-        );
-
-        assert_eq!(
-            other_obj_list
-                .objects
-                .values()
-                .next()
-                .unwrap()
-                .0
-                .get_state(),
-            crate::analysis::pointer_inference::object::ObjectState::Alive
-        );
-        other_obj_list
-            .mark_mem_object_as_freed(&modified_heap_pointer)
-            .unwrap();
-        assert_eq!(
-            other_obj_list
-                .objects
-                .values()
-                .next()
-                .unwrap()
-                .0
-                .get_state(),
-            crate::analysis::pointer_inference::object::ObjectState::Dangling
-        );
-    }
-
-    #[test]
-    fn append_unknown_objects_test() {
-        let mut obj_list = AbstractObjectList::from_stack_id(new_id("stack"), ByteSize::new(8));
-
-        let mut other_obj_list =
-            AbstractObjectList::from_stack_id(new_id("stack"), ByteSize::new(8));
-        other_obj_list.add_abstract_object(
-            new_id("heap_obj"),
-            bv(0).into(),
-            ObjectType::Heap,
-            ByteSize::new(8),
-        );
-
-        obj_list.append_unknown_objects(&other_obj_list);
-        assert_eq!(obj_list.objects.len(), 2);
-        assert!(obj_list.objects.get(&new_id("stack")).is_some());
-        assert!(obj_list.objects.get(&new_id("heap_obj")).is_some());
-    }
-}
+mod tests;
